requirements.txt
click==8.1.3
Flask==2.2.2
Flask-MySQLdb==1.0.1
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysql-connector==2.2.9
mysql-connector-python==8.0.31
mysqlclient==2.1.1
paho-mqtt==1.6.1
protobuf==3.20.1
Werkzeug==2.2.2
zipp==3.10.0


'{"data":["{\"node\":0,\"temp\":22.79999924,\"hum\":65,\"id\":317221743}","{\"node\":1,\"temp\":22.10000038,\"hum\":65,\"id\":317221787}"],"avg_temp":20.1738, "avg_hum":60,"delay":10}'

